// packages/plugin-inbox-ui/.erxes/plugin-src/inbox/types.ts
export interface IVideoCallData {
  url: string;
  name?: string;
  status?: string;
  recordingLinks?: string[];
}

export interface IConversation {
  _id: string;
  content?: string;
  integrationId: string;
  customerId?: string;
  userId?: string;
  assignedUserId?: string;
  participatedUserIds?: string[];
  readUserIds?: string[];
  createdAt: Date;

  closedAt?: Date;
  closedUserId?: string;

  status?: string;
  messageCount?: number;
  tagIds?: string[];

  // number of total conversations
  number?: number;

  integration: IIntegration;
  customer: ICustomer;
  assignedUser: IUser;
  participatedUsers?: IUser[];
  tags: ITag[];
  updatedAt: Date;
  idleTime: number;
  facebookPost?: IFacebookPost;
  callProAudio?: string;
  videoCallData?: IVideoCallData;

  isFacebookTaggedMessage?: boolean;

  customFieldsData?: {
    [key: string]: any;
  };

  bookingProductId?: string;
}

interface IEngageDataRules {
  kind: string;
  text: string;
  condition: string;
  value?: string;
}

export interface IFacebookPost {
  postId: string;
  recipientId: string;
  senderId: string;
  content: string;
  erxesApiId?: string;
  attachments: string[];
  timestamp: Date;
  permalink_url: string;
}
// ...
export type CustomerDetailQueryResponse = {
  customerDetail: ICustomer;
  loading: boolean;
};